requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
fastapi==0.115.6
h11==0.14.0
idna==3.10
joblib==1.4.2
numpy==2.2.1
pandas==2.2.3
pydantic==2.10.4
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
requests==2.32.3
scikit-learn==1.6.0
scipy==1.14.1
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
